Update README.md
All good. 
There was a small issue with the URL as the } got inside the URL.
I suggest not putting the path inside the commands as this prevents copy-pasting the code.
I've included the submodule update (not sure if always necessary)

I've followed the way commands are written here: 
https://github.com/5G-MAG/Documentation-and-Architecture/wiki/MBMS-Middleware
I've added the ~ in all commands as it may not be clear where to execute everything while enabling the copy-paste... but no strong view.